- Corrected an error that generates the mask with float values instead of integers.

+++ staples_detection/base/result/mask_detection_result.py
@@ -8,9 +8,9 @@ class MaskDetectionResult:
                  colormask: numpy.ndarray,
                  elapsed_time: float):
 
-        self.final_binary_mask = final_mask
+        self.final_binary_mask = final_mask.astype(int)
         self.colormask = colormask
         self.elapsed_time = round(elapsed_time, 5)
 
-        self.final_mask = numpy.zeros_like(self.final_binary_mask)
+        self.final_mask = numpy.zeros_like(self.final_binary_mask).astype(int)
         self.final_mask[self.final_binary_mask == True] = 255
